Refactor to directory
