--- NavgationTitlePoppingDemo/Extensions/UIColor+Title.swift
@@ -1,18 +1,0 @@
-//
-//  Created by lili on 2018/5/17.
-//
-//  Copyright © 2018年 fun. All rights reserved.
-//
-
-import UIKit
-
-extension UIColor {
-    
-    public static var nav_purple: UIColor {
-        return UIColor(red: 124.0 / 255, green: 79.0 / 255, blue: 247.0 / 255, alpha: 1.0)
-    }
-    
-    public static var nav_black: UIColor {
-        return UIColor(red: 55.0 / 255, green: 55.0 / 255, blue: 55.0 / 255, alpha: 1.0)
-    }
-}
