--- NavgationTitlePoppingDemo/Extensions/UIFont+Title.swift
@@ -1,17 +1,0 @@
-//
-//  Created by lili on 2018/5/17.
-//
-//  Copyright © 2018年 fun. All rights reserved.
-//
-
-import UIKit
-
-extension UIFont {
-    public static var nav_regular: UIFont {
-        return .systemFont(ofSize: 18, weight: .medium)
-    }
-    
-    public static var nav_small: UIFont {
-        return .systemFont(ofSize: 13, weight: .medium)
-    }
-}
